% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GWSPIA.r
\name{spiaDBS}
\alias{spiaDBS}
\title{done GWSPIA function to return a data.frame by using spiaDBS}
\usage{
spiaDBS(de = NULL, all = NULL, allt = NULL, betweennesslist = NULL,
  DEdegree = NULL, organism = "hsa", data.dir = NULL, pathids = NULL,
  nB = 2000, plots = FALSE, verbose = TRUE, beta = NULL,
  combine = "fisher")
}
\arguments{
\item{de, should}{be a vector}

\item{all, should}{be a vector}

\item{allt, should}{be a vector}

\item{betweennesslist, should}{be a list}

\item{DEdegree, should}{be a vector}
}
\description{
done GWSPIA function to return a data.frame by using spiaDBS
}
